Did a little bit of work
Separated the client and the Linda server in different packages, and created the tuple class and coded things in it(idk if it works ) and the server and  client should work as inteded(IE connect to one another and create as many threads of linda as clients connected ).
I just want to die.