Almost done, just needs a few tweaks!
What is left is to make a check to see if the server1_3 is running, if not the replica should take over and respond instead of the serv1_3, and to see if the remove function works as intended or not.
One more thing to keep in mind is to check if the serv1_3 has only one "out" or more, because it can lag the servers.